Support #blpop.
If you end up calling #blpop with a timeout of 0 and there's no data
available, then an exception will be raised. This seemed preferable to
just waiting forever. After all, this is an in-memory mock; nobody is
going to come along and rpush some data into any of the lists you're
blocking on.