Support #setbit (kind of).
There are probably corner cases that it doesn't handle. Redis is
written in C++, so it can just treat a string as a byte array, but in
Ruby, we don't have that luxury. Sure, there's String#each_byte, but
if you turn a character into a byte array and twiddle its bits, then
that byte array might not convert back to a valid character.

Patches welcome. :)