Support #expire, #expireat, and #persist.
To get this to work, MockRedis no longer stores data. Instead, it has
a MockRedis::DataStore object that actually stores the data, and
MockRedis forwards messages (via #method_missing) to the DataStore,
but before it sends each message, it tells the DataStore to expire its
data.

Performance will probably suffer if you've got large numbers of keys
with expirations on them due to the naive way in which key expirations
are handled (lots of O(N) array operations). Patches welcome. :)