Small spec tweak.
The test for #multi(&blk) was running just against the mock because
calling methods on @redises inside the block made things all
wonky. However, the block gets an argument that's the Redis object on
which multi is called, so if the test uses that instead, then there's
no confusion around the multiplexer.