Fix rename spec
This spec was failing because in 'real' redis, e.g. the part of the test
that hits an actual redis-server, the key 'foo' was set.

There's already a before block in spec_helper that clears out all keys
in the mock-redis-test:* namespace, so we changed the spec to use a key
under this namespace, which we know will not exist in either mock or
real redis.

Change-Id: I5ae9d1454034ee87f96724dd6c3597ef049cda9e
Reviewed-on: https://gerrit.causes.com/11458
Reviewed-by: Elliot Block <elliot.block@causes.com>
Tested-by: Elliot Block <elliot.block@causes.com>