Add support for count in srandmember
The `count` argument was present but not used in the `srandmember`
method.

This change adds support for the `count` argument based on the
specification of `srandmember` as described on:

http://redis.io/commands/srandmember

We can't use result checking between MockRedis and Redis in the specs
because `srandmember` gives random members so we'll likely get
different results between MockRedis and Redis.

Change-Id: I639cd5ff36b3e4b1b5a5c80a50e1c261ba131b0b
Reviewed-on: https://gerrit.causes.com/33202
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>