Fix error messages for keys with wrong value type
In preparation for upgrading the version of Redis used in tests to 2.8,
it was noticed that the error message for incorrect value types of
various commands changed its prefix from 'ERR' to 'WRONGTYPE'. This is
confirmed by the release notes here:

https://raw.github.com/antirez/redis/2.8/00-RELEASENOTES

Fix the broken specs by updating the error message prefix.

Change-Id: Iea3fe537a475e9783edf6487b2cd312e0be5bb53
Reviewed-on: https://gerrit.causes.com/33196
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>