Update TTL behavior when no key exists
In the process of moving the test suite onto Redis 2.8, change the
behavior of the `ttl` command to return -2 when the key does not exist.

From the docs at http://redis.io/commands/ttl:

Starting with Redis 2.8 the return value in case of error changed:

* The command returns -2 if the key does not exist.
* The command returns -1 if the key exists but has no associated expire.

This modifies the implementation and the associated spec to conform to
this new behavior.

Change-Id: I94b877aa4cc71b9454bc6a3f95b8185d11df505b
Reviewed-on: https://gerrit.causes.com/33197
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>