Round start time in pexpire specs
We were seeing this test spuriously fail under JRuby 1.7.9. After some
investigation, it's clear that it has something to do with differing
precisions in the underlying objects (i.e. Java Time versus Ruby Time)
and how they are rounded.

Since this problem only occurs in JRuby and only potentially breaks
functionality at the millisecond level (e.g. the expire will be off by
+/- 1 millisecond), I'm not inclined to dig any deeper.

A simple solution is to round the starting time so that the time is
always guaranteed to be at the start of a second.

Change-Id: Ib7280cba69708afd2c71bc8b78c6a28ea287a807
Reviewed-on: https://gerrit.causes.com/33199
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>