Add support to del for accepting an array of keys
There was a discrepancy between the Redis gem and MockRedis when
deleting arrays of keys. Consider the following:

Redis

    irb(main):001:0> $redis.set 'a','a'
    => "OK"
    irb(main):002:0> $redis.set 'b','b'
    => "OK"
    irb(main):003:0> $redis.del ['a','b']
    => 2

MockRedis

    irb(main):001:0> $mock_redis.set 'a','a'
    => "OK"
    irb(main):002:0> $mock_redis.set 'b','b'
    => "OK"
    irb(main):003:0> $mock_redis.del ['a','b']
    => 0

The issue is that MockRedis expected the keys given to `del` to be
provided as separate arguments, not an array.

Solve by flattening the arguments array and adding a test.

Fixes #65

Change-Id: I0cf9044c37eb6e615946af2c2e533afe8d475f07
Reviewed-on: https://gerrit.causes.com/33216
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>