Don't check that a specific error wasn't raised
After upgrading RSpec the test suite was reporting deprecating warnings
for the few places where we checks that a specific error _was not_
raised.

The problem with these specs is that if a different error is raised,
they will likely erroneously pass, as it is typically expected that _no_
error is raised.

Replace these uses with simple `should_not raise_error` to fix the
deprecation warning.

In the process, two tests were discovered to have false positives. The
first had an incorrect variable, and the second was just wrong since
`brpoplpush` does not accept subsecond timeout values (according to the
docs the timeout is always interpreted as an integer).

Change-Id: Ie37429ab5f2225d55ec90533f963113e24873151
Reviewed-on: http://gerrit.causes.com/37003
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>