feat: Implement getdel command (#250)
This adds support to [GETDEL](https://redis.io/commands/getdel/)
introduced in Redis 6.2.0 (implemented on commit
https://github.com/redis/redis-rb/commit/94d606c04c0322600cf1c4dfe4ebba12155e7ea5)

I initially implemented it as

```ruby
    def getdel(key)
      get(key).tap { del(key) }
    end
```

but I couldn't find `tap` usages on the codebase, so I preferred to
introduce a variable, and return it after the `del` call. Please do let
me know in case there's a standard to follow.

Thanks!