RSpec: use expect assertion syntax (#258)
Back in RSpec 2.11, a new expectation syntax was introduced. Check out
the blog post that outlines the shortcomings of the `should` syntax and
how these are resolved with the new `expect` syntax.

https://rspec.info/blog/2012/06/rspecs-new-expectation-syntax/

I propose the RSpec test suite be migrated to this `expect` syntax.
Luckily this is quite easy as there is a conversion tool that does all
the heavy lifting: [transpec](http://yujinakayama.me/transpec/).

I ran this tool across the test suite and this pull request consists of
the changes.