Add support for scanning on keys containing hash tags (#273)
This PR adds support for scanning on keys containing hash tags. This is
mainly relevant when working with Redis in cluster mode. If `mock_redis`
is not supposed to support Cluster mode, I understand if you want to
close this PR.

[Hash tags](https://redis.io/docs/management/scaling/) are parts of a
key surrounded by `{}`. This is used in cluster mode to ensure a set of
keys is stored on the same node. MockRedis does not support this and
scanning on a key containing a hash tag does not work because the
resulting regex pattern created by `redis_pattern_to_ruby_regex` does
not escape curly braces.

The added commit is a suggestion on how support can be added. If there
is a better way, feel free to edit directly or notify me and I will make
the required change.