Real Redis adds an Array as a stringified member (#284)
We were flattening the wrapped array here to get rid of any nesting of
arrays we might have accidentally done by wrapping with an Array literal
(`[]`). Instead, `Kernel#Array` will leave an existing Array alone, and
wrap any single object in an Array. This is consistent with real Redis'
behavior.