Fix `Database.call` with expire methods (#334)
# What?
* It fixes calls to `expire*` methods when the calling happens through
the `.call` method of MockRedis::Database
* It fixes a test that will inconsistently fail when we run tests one
after another, within a few seconds of each run

# Why?
`expire` values are not respected when we set these via the Database
object. So tests that depend on testing that we set the ttl correctly
will fail. This happens because the `.call` method fails to call
successfully `public_send` due to the need to parse out the kwarg
arguments.

# How?
The fix was to identify if we have an `expire` call in the command. If
that is the case, we call a `send_expire` method that will correctly
transform the call so that we can call those methods with the correct
format.

For the test, I noticed that adding a slight wait for 1 millisecond was
enough to consistently get that test passed

# Notes

I didn't add a version number because I didn't want to be be arrogant
and assume that the PR will be passed. I am happy to modify things or
try some other solution if you prefer that.

Thanks so much for such a useful gem